chore: 여러 개의 atom 파일 하나로 합침
