--- week7/src/atom/headerAtom.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const animateState = atom({
-  key: "animateState",
-  default: false,
-});
